🐛🤖 Normalize staging branch names and support --userId (Codex review)
- Extract the branch->staging-hostname normalization into a shared
  devTools/stagingHostname.ts and use it in both query.ts and
  callAdminApi.ts, instead of callAdminApi.ts interpolating the raw
  branch name. Fixes staging hosts for branches with slashes, dots,
  underscores, or names over 28 characters.
- Add --userId, sent as x-act-as-user, so set/unset can attribute
  writes to the actual editor instead of always the shared
  ADMIN_API_KEY's own owner.
